handle some bug & improve performance
