+++ src/app/components/encrypt-password/encrypt-password.scss
@@ -125,7 +125,7 @@ encrypt-password {
       }
       .footer {
         height: 76px;
-        margin-bottom: 1rem;
+        margin-bottom: 1.5rem;
         align-self: end;
       }
     }
